Fix same TV Show names with different case
Before this commit the mapper did not adjust the case of the TV Show
name. As a consequence, in the case sensitive file systems (such as the
ones used on linux) the organizer would create multiple folders for the
same TV Show: one for each casing that it found.

After this commit, the organizer will always create the same folder with
the exact same case regardless of the casing used in the episode name.